Move delegate package to triplea main package
