+++ src/games/strategy/engine/xml/TestDelegate.java
@@ -2,7 +2,7 @@ package games.strategy.engine.xml;
 
 import java.io.Serializable;
 
-import games.strategy.common.delegate.AbstractDelegate;
+import games.strategy.triplea.delegate.AbstractDelegate;
 import games.strategy.engine.message.IRemote;
 
 /**
