+++ src/games/strategy/triplea/delegate/BasePersistentDelegate.java
@@ -1,4 +1,4 @@
-package games.strategy.common.delegate;
+package games.strategy.triplea.delegate;
 
 import games.strategy.engine.delegate.IPersistentDelegate;
