+++ src/games/strategy/triplea/delegate/GameDelegateHistoryWriter.java
@@ -1,4 +1,4 @@
-package games.strategy.common.delegate;
+package games.strategy.triplea.delegate;
 
 import games.strategy.engine.data.GameData;
 import games.strategy.engine.history.IDelegateHistoryWriter;
