+++ src/games/strategy/triplea/delegate/RandomStartDelegate.java
@@ -9,7 +9,6 @@ import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
 
-import games.strategy.common.delegate.BaseTripleADelegate;
 import games.strategy.engine.data.ChangeFactory;
 import games.strategy.engine.data.CompositeChange;
 import games.strategy.engine.data.GameData;
