+++ src/games/strategy/triplea/delegate/SpecialMoveDelegate.java
@@ -7,8 +7,6 @@ import java.util.List;
 import java.util.Map;
 import java.util.Set;
 
-import games.strategy.common.delegate.BaseEditDelegate;
-import games.strategy.common.delegate.GameDelegateBridge;
 import games.strategy.engine.data.Change;
 import games.strategy.engine.data.ChangeFactory;
 import games.strategy.engine.data.CompositeChange;
