+++ src/games/strategy/triplea/delegate/StrategicBombingRaidBattle.java
@@ -9,7 +9,6 @@ import java.util.Iterator;
 import java.util.List;
 import java.util.Set;
 
-import games.strategy.common.delegate.BaseEditDelegate;
 import games.strategy.engine.data.Change;
 import games.strategy.engine.data.ChangeFactory;
 import games.strategy.engine.data.GameData;
