+++ src/games/strategy/triplea/delegate/TechnologyDelegate.java
@@ -11,9 +11,6 @@ import java.util.List;
 import java.util.Map;
 import java.util.Map.Entry;
 
-import games.strategy.common.delegate.BaseEditDelegate;
-import games.strategy.common.delegate.BaseTripleADelegate;
-import games.strategy.common.delegate.GameDelegateBridge;
 import games.strategy.engine.data.Change;
 import games.strategy.engine.data.ChangeFactory;
 import games.strategy.engine.data.GameData;
