+++ src/games/strategy/triplea/ui/MovePanel.java
@@ -22,7 +22,7 @@ import javax.swing.AbstractAction;
 import javax.swing.Action;
 import javax.swing.JOptionPane;
 
-import games.strategy.common.delegate.BaseEditDelegate;
+import games.strategy.triplea.delegate.BaseEditDelegate;
 import games.strategy.debug.ClientLogger;
 import games.strategy.engine.data.GameData;
 import games.strategy.engine.data.PlayerID;
